A working approach.
thumb2.py can do everything arm9dis.py could do, and more.
It knows a handful of opcodes, and i think that the rest could
be added fairly easily.

I wanted to avoid a big old switch statement, so with the exception of
the BL and BLX instructions (which are 32 bytes, requiring special
handling), opcodes are implemented as classes with methods for
matching and parsing the opcodes.